+++ laser_line_extraction/include/laser_line_extraction/utilities.h
