+++ laser_line_extraction_interfaces/msg/LineSegment.msg
@@ -1,0 +1,5 @@
+float64 radius
+float64 angle
+float64[4] covariance
+float64[2] start
+float64[2] end
