+++ laser_line_extraction_interfaces/msg/LineSegmentList.msg
@@ -1,0 +1,2 @@
+std_msgs/Header header
+LineSegment[] line_segments
