--- launch/example.launch
@@ -1,20 +1,0 @@
-<launch>
-  <node name="line_extractor" pkg="laser_line_extraction" type="line_extraction_node">
-    <param name="~frequency" value="30.0" />
-    <param name="~frame_id" value="laser" />
-    <param name="~scan_topic" value="scan" />
-    <param name="~publish_markers" value="true" />
-    <param name="~bearing_std_dev" value="1e-5" />
-    <param name="~range_std_dev" value="0.012" />
-    <param name="~least_sq_angle_thresh" value="0.0001" />
-    <param name="~least_sq_radius_thresh" value="0.0001" />
-    <param name="~max_line_gap" value="0.5" />
-    <param name="~min_line_length" value="0.7" />
-    <param name="~min_range" value="0.5" />
-    <param name="~max_range" value="250.0" />
-    <param name="~min_split_dist" value="0.04" />
-    <param name="~outlier_dist" value="0.06" />
-    <param name="~min_line_points" value="10" />
-  </node>
-</launch>
-
