--- msg/LineSegmentList.msg
@@ -1,2 +1,0 @@
-Header header
-LineSegment[] line_segments
