--- package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>laser_line_extraction</name>
-  <version>0.1.0</version>
-  <description>
-      A ROS package to extract line segments from LaserScan messages.
-  </description>
-  <maintainer email="m.gallant@queensu.ca">Marc Gallant</maintainer>
-  <license>BSD</license>
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>cmake_modules</build_depend>
-  <build_depend>eigen</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>message_generation</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>rospy</build_depend>
-  <build_depend>sensor_msgs</build_depend>
-  <build_depend>visualization_msgs</build_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>message_runtime</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>rospy</run_depend>
-  <run_depend>sensor_msgs</run_depend>
-  <run_depend>visualization_msgs</run_depend>
-</package>
